add kdoc to paging source and move impl to its own file
